+++ extensions/google-calendar/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Google Calendar Changelog
 
+## [1.3.2] - 2025-12-01
+
+- Feat(create-event): Added All-day Event support to the Create Event form.
+- UI: Introduced new "Event Time" selector for switching between timed and all-day events.
+
 ## [1.3.1] - 2025-11-25
 
 - Allow the user to configure if they wish to open a meeting directly as the default action instead of the calendar event, defaults to the existing behaviour.